Added types, get it now functionality for activating popup only
